Split out util module
