+++ bouncer/views.py
@@ -1,68 +1,16 @@
 import json
 from urllib import parse
 
-import elasticsearch
 from elasticsearch import exceptions
 from pyramid import httpexceptions
 from pyramid import i18n
 from pyramid import view
 from statsd.defaults.env import statsd
 
+from bouncer import util
+
 
 _ = i18n.TranslationStringFactory(__package__)
-
-
-class InvalidAnnotationError(Exception):
-
-    """Raised if an annotation from Elasticsearch can't be parsed."""
-
-    def __init__(self, message, reason):
-        """
-        Return a new InvalidAnnotationError instance.
-
-        :param message: a user-friendly error message
-        :type message: string
-
-        :param reason: a computer-friendly unique string identifying the reason
-            the exception was raised
-        :type reason: string
-
-        """
-        self.message = message
-        self.reason = reason
-
-    def __str__(self):
-        return self.message
-
-
-def parse_document(document):
-    """
-    Return the ID and URI from the given Elasticsearch annotation document.
-
-    Return the annotation ID and the annotated document's URI from the given
-    Elasticsearch annotation document.
-
-    :param document: the Elasticsearch annotation document to parse
-    :type document: dict
-
-    :rtype: 2-tuple of annotation ID (string) and document URI (string)
-
-    """
-    # We assume that Elasticsearch documents always have "_id" and "_source".
-    annotation_id = document["_id"]
-    annotation = document["_source"]
-
-    try:
-        document_uri = annotation["uri"]
-    except KeyError:
-        raise InvalidAnnotationError(
-            _("The annotation has no URI"), "annotation_has_no_uri")
-
-    if not isinstance(document_uri, str):
-        raise InvalidAnnotationError(
-            _("The annotation has an invalid document URI"), "uri_not_a_string")
-
-    return (annotation_id, document_uri)
 
 
 @view.view_defaults(renderer="bouncer:templates/annotation.html.jinja2")
@@ -77,7 +25,7 @@ class AnnotationController(object):
         settings = self.request.registry.settings
 
         try:
-            document = elasticsearch_client(settings).get(
+            document = util.elasticsearch_client(settings).get(
                 index=settings["elasticsearch_index"],
                 doc_type="annotation",
                 id=self.request.matchdict["id"])
@@ -86,8 +34,8 @@ class AnnotationController(object):
             raise httpexceptions.HTTPNotFound(_("Annotation not found"))
 
         try:
-            annotation_id, document_uri = parse_document(document)
-        except InvalidAnnotationError as exc:
+            annotation_id, document_uri = util.parse_document(document)
+        except util.InvalidAnnotationError as exc:
             statsd.incr("views.annotation.422.{}".format(exc.reason))
             raise httpexceptions.HTTPUnprocessableEntity(str(exc))
 
@@ -121,13 +69,6 @@ class AnnotationController(object):
                 "extensionUrl": extension_url,
             })
         }
-
-
-def elasticsearch_client(settings):
-    return elasticsearch.Elasticsearch(
-        host=settings["elasticsearch_host"],
-        port=settings["elasticsearch_port"],
-    )
 
 
 @view.view_config(renderer="bouncer:templates/index.html.jinja2",
